requirements.txt
accelerate==0.32.1
av==12.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cmake==3.30.0
contourpy==1.2.1
cycler==0.12.1
decorator==4.4.2
decord==0.6.0
draccus==0.8.0
einops==0.8.0
filelock==3.15.4
# flash-attn==2.5.9.post1
fonttools==4.53.1
fsspec==2024.6.1
fvcore==0.1.5.post20221221
huggingface-hub==0.23.4
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
importlib_resources==6.4.0
iopath==0.1.10
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
lit==18.1.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
mergedeep==1.3.4
moviepy==1.0.3
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
parameterized==0.9.0
peft==0.3.0
pillow==10.4.0
portalocker==2.10.0
proglog==0.1.10
protobuf==5.27.2
psutil==6.0.0
# pycocoevalcap @ git+https://github.com/salaniz/pycocoevalcap.git@a24f74c408c918f1f4ec34e9514bc8a76ce41ffd
pycocotools==2.0.8
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytorchvideo==0.1.5
pytz==2024.1
PyYAML==6.0.1
pyyaml-include==1.4.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
safetensors==0.4.3
sentencepiece==0.2.0
six==1.16.0
sympy==1.12.1
tabulate==0.9.0
termcolor==2.4.0
timm==1.0.7
tokenizers==0.19.1
toml==0.10.2
torch==2.2.1
torchaudio==2.2.1
torchvision==0.17.1
tqdm==4.66.4
transformers==4.40.1
triton==2.2.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
yacs==0.1.8
zipp==3.19.2
